forwardauth: allow config type of handler to run
