+++ config/options_test.go
@@ -44,6 +44,12 @@ func Test_Validate(t *testing.T) {
 	invalidStorageType.DataBrokerStorageType = "foo"
 	missingStorageDSN := testOptions()
 	missingStorageDSN.DataBrokerStorageType = "redis"
+	invalidForwardAuthType := testOptions()
+	invalidForwardAuthType.ForwardAuthURLString = "https://forwardauth.example.com"
+	invalidForwardAuthType.ForwardAuthType = "invalid"
+	goodForwardAuth := testOptions()
+	goodForwardAuth.ForwardAuthURLString = "https://forwardauth.example.com"
+	goodForwardAuth.ForwardAuthType = ProxyTypeTraefik
 
 	tests := []struct {
 		name     string
@@ -57,6 +63,8 @@ func Test_Validate(t *testing.T) {
 		{"policy file specified", badPolicyFile, true},
 		{"invalid databroker storage type", invalidStorageType, true},
 		{"missing databroker storage dsn", missingStorageDSN, true},
+		{"good forward auth", goodForwardAuth, false},
+		{"invalid forward auth type", invalidForwardAuthType, true},
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
@@ -243,6 +251,7 @@ func TestOptionsFromViper(t *testing.T) {
 				RefreshDirectoryTimeout:  1 * time.Minute,
 				RefreshDirectoryInterval: 10 * time.Minute,
 				QPS:                      1.0,
+				ForwardAuthType:          ProxyTypeNginx,
 				DataBrokerStorageType:    "memory",
 			},
 			false},
@@ -261,6 +270,7 @@ func TestOptionsFromViper(t *testing.T) {
 				RefreshDirectoryTimeout:         1 * time.Minute,
 				RefreshDirectoryInterval:        10 * time.Minute,
 				QPS:                             1.0,
+				ForwardAuthType:                 ProxyTypeNginx,
 				DataBrokerStorageType:           "memory",
 			},
 			false},
