+++ forwardauth/handlers.go
@@ -7,11 +7,18 @@ import (
 	"net/http"
 	"net/url"
 
+	"github.com/gorilla/mux"
+
 	"github.com/pomerium/pomerium/internal/httputil"
 	"github.com/pomerium/pomerium/internal/sessions"
 	"github.com/pomerium/pomerium/internal/urlutil"
 	"github.com/pomerium/pomerium/pkg/cryptutil"
 )
+
+var registerHandler = map[string]func(*ForwardAuth, *mux.Router){
+	ProxyTypeNginx:   registerNginxHandlers,
+	ProxyTypeTraefik: registerTraefikHandlers,
+}
 
 // saveCallbackSession takes an encrypted per-route session token, and decrypts
 // it using the shared service key, then stores it the local session store.
@@ -47,72 +54,9 @@ func (fa *ForwardAuth) registerFwdAuthHandlers() http.Handler {
 	})
 	r.Use(fa.jwtClaimMiddleware)
 
-	// NGNIX's forward-auth capabilities are split across two settings:
-	// `auth-url` and `auth-signin` which correspond to `verify` and `auth-url`
-	//
-	// NOTE: Route order matters here which makes the request flow confusing
-	// 		 to reason about so each step has a postfix order step.
-
-	// nginx 3: save the returned session post authenticate flow
-	r.Handle("/verify", httputil.HandlerFunc(fa.nginxCallback)).
-		Queries("uri", "{uri}", urlutil.QuerySessionEncrypted, "", urlutil.QueryRedirectURI, "")
-
-	// nginx 1: verify. Return 401 if invalid and NGINX will call `auth-signin`
-	r.Handle("/verify", fa.Verify(true)).Queries("uri", "{uri}")
-
-	// nginx 4: redirect the user back to their originally requested location.
-	r.Handle("/", httputil.HandlerFunc(fa.nginxPostCallbackRedirect)).
-		Queries("uri", "{uri}", urlutil.QuerySessionEncrypted, "", urlutil.QueryRedirectURI, "")
-
-	// traefik 2: save the returned session post authenticate flow
-	r.Handle("/", httputil.HandlerFunc(fa.forwardedURIHeaderCallback)).
-		HeadersRegexp(httputil.HeaderForwardedURI, urlutil.QuerySessionEncrypted)
-
-	// nginx 2 / traefik 1: verify and then start authenticate flow
-	r.Handle("/", fa.Verify(false))
-
+	proxyType := fa.state.Load().proxyType
+	registerHandler[proxyType](fa, r)
 	return r
-}
-
-// nginxPostCallbackRedirect redirects the user to their original destination
-// in order to drop the authenticate related query params
-func (fa *ForwardAuth) nginxPostCallbackRedirect(w http.ResponseWriter, r *http.Request) error {
-	w.Header().Set("Content-Type", "text/plain; charset=utf-8")
-	httputil.Redirect(w, r, r.FormValue(urlutil.QueryRedirectURI), http.StatusFound)
-	return nil
-}
-
-// nginxCallback saves the returned session post callback and then returns an
-// unauthorized status in order to restart the request flow process. Strangely
-// we need to throw a 401 after saving the session to redirect the user
-// to their originally desired location.
-func (fa *ForwardAuth) nginxCallback(w http.ResponseWriter, r *http.Request) error {
-	encryptedSession := r.FormValue(urlutil.QuerySessionEncrypted)
-	if _, err := fa.saveCallbackSession(w, r, encryptedSession); err != nil {
-		return httputil.NewError(http.StatusBadRequest, err)
-	}
-	w.Header().Set("Content-Type", "text/plain; charset=utf-8")
-	w.WriteHeader(http.StatusUnauthorized)
-	return nil
-}
-
-// forwardedURIHeaderCallback handles the post-authentication callback from
-// forwarding proxies that support the `X-Forwarded-Uri`.
-func (fa *ForwardAuth) forwardedURIHeaderCallback(w http.ResponseWriter, r *http.Request) error {
-	forwardedURL, err := url.Parse(r.Header.Get(httputil.HeaderForwardedURI))
-	if err != nil {
-		return httputil.NewError(http.StatusBadRequest, err)
-	}
-	q := forwardedURL.Query()
-	redirectURLString := q.Get(urlutil.QueryRedirectURI)
-	encryptedSession := q.Get(urlutil.QuerySessionEncrypted)
-
-	if _, err := fa.saveCallbackSession(w, r, encryptedSession); err != nil {
-		return httputil.NewError(http.StatusBadRequest, err)
-	}
-	w.Header().Set("Content-Type", "text/plain; charset=utf-8")
-	httputil.Redirect(w, r, redirectURLString, http.StatusFound)
-	return nil
 }
 
 // Verify checks a user's credentials for an arbitrary host. If the user
@@ -161,14 +105,14 @@ func (fa *ForwardAuth) Verify(verifyOnly bool) http.Handler {
 			return httputil.NewError(http.StatusUnauthorized, err)
 		}
 
-		fa.forwardAuthRedirectToSignInWithURI(w, r, uri)
+		fa.redirectToSignInWithURI(w, r, uri)
 		return nil
 	})
 }
 
-// forwardAuthRedirectToSignInWithURI redirects request to authenticate signin url,
+// redirectToSignInWithURI redirects request to authenticate signin url,
 // with all necessary information extracted from given input uri.
-func (fa *ForwardAuth) forwardAuthRedirectToSignInWithURI(w http.ResponseWriter, r *http.Request, uri *url.URL) {
+func (fa *ForwardAuth) redirectToSignInWithURI(w http.ResponseWriter, r *http.Request, uri *url.URL) {
 	state := fa.state.Load()
 
 	// Traefik set the uri in the header, we must set it in redirect uri if present. Otherwise, request like
